Fix to replace from useWheelPicker function to  useObserver function
